requirements.txt
allure-pytest==2.13.5
selenium==4.20.0
pytest==8.2.0
pytest-xdist==3.5.0
requests==2.31.0
faker==26.1.0
webdriver-manager==3.8.6
pylint==3.0.3
jsonschema==4.21.1